+++ docs/linkedlist/MergeSortLinkedList.md
@@ -1,26 +1,6 @@
-<!----- Conversion time: 0.637 seconds.
-
-
-Using this Markdown file:
-
-1. Cut and paste this output into your source file.
-2. See the notes and action items below regarding this conversion run.
-3. Check the rendered output (headings, lists, code blocks, tables) for proper
-   formatting and use a linkchecker before you publish this page.
-
-Conversion notes:
-
-* Docs to Markdown version 1.0β14
-* Mon Feb 11 2019 00:37:06 GMT-0800 (PST)
-* Source doc: https://docs.google.com/open?id=1jSNfGGIswI8KmR2GnvrvVqk20WfGpFnn5ysf5Ce0tuQ
------>
-
-
-
 # Merge Sort Linked List
 
 [https://app.laicode.io/app/problem/29](https://app.laicode.io/app/problem/29)
-
 
 ## Description
 
@@ -28,12 +8,10 @@ Given a singly-linked list, where each node contains an integer value, sort it i
 
 Examples
 
-
-
-*   null, is sorted to null
-*   1 -> null, is sorted to 1 -> null
-*   1 -> 2 -> 3 -> null, is sorted to 1 -> 2 -> 3 -> null
-*   4 -> 2 -> 6 -> -3 -> 5 -> null, is sorted to -3 -> 2 -> 4 -> 5 -> 6
+- null, is sorted to null
+- 1 -> null, is sorted to 1 -> null
+- 1 -> 2 -> 3 -> null, is sorted to 1 -> 2 -> 3 -> null
+- 4 -> 2 -> 6 -> -3 -> 5 -> null, is sorted to -3 -> 2 -> 4 -> 5 -> 6
 
 Medium
 
@@ -43,22 +21,17 @@ Sort
 
 Two Pointers
 
-
 ## Assumption
 
 The linked list is singly-linked and should not be null or empty. Nor should it have cycles in it.
-
 
 ## Algorithm
 
 The merge sort algorithm can be perfectly fit to solve this problem. The basic idea in a high level point of view is to divide the linked list into two parts each time and recursively sort the two parts. After the divided linked lists are sorted accordingly, we merge them together based on the principle of "move the smaller one".
 
-
 ## Solution
 
-
 ### Code
-
 
 ```java
 /**
@@ -72,6 +45,7 @@ The merge sort algorithm can be perfectly fit to solve this problem. The basic i
  * }
  */
 public class Solution {
+
   public ListNode mergeSort(ListNode head) {
     // Write your solution here
     // Corner cases
@@ -136,19 +110,12 @@ public class Solution {
 }
 ```
 
-
-
 ### Complexity
-
 
 #### Time
 
 The linked lists are divided by ½ each time, so there are log(n) levels in the call stack. In each level, finding the middle node of the list takes O(n) time. So does the merge process. Therefore, the total time complexity is O(n log(n)).
 
-
 #### Space
 
 Because there are log(n) levels in the recursion tree/call stack, and only variables/objects with constant spaces are created in each level. The total space complexity is O(log(n)).
-
-
-<!-- Docs to Markdown version 1.0β14 -->
